Compilation Docker images and services finalized
With this update we get rid of the single-use images for a more flexible command based interface.

This new approach assumes that the user is going to compile the benchmarks provided under the `applications` folder.

The images can now be reused as they only provide access to the compiler and are not bound to a specific benchmark.